unit tests not supporting abstract class
